Moving the File Writer to its own Plugin Module.
Refactoring Unit Tests.

Got rid of Legacy Logger.
