+++ service/local-storage-plugin/src/main/java/com/solace/maas/ep/runtime/agent/plugin/localstorage/LocalStorageApplication.java
@@ -1,0 +1,11 @@
+package com.solace.maas.ep.runtime.agent.plugin.localstorage;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class LocalStorageApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(LocalStorageApplication.class, args);
+    }
+}
