+++ service/local-storage-plugin/src/main/java/com/solace/maas/ep/runtime/agent/plugin/localstorage/route/delegate/DataCollectionFileWriterDelegateImpl.java
@@ -1,0 +1,21 @@
+package com.solace.maas.ep.runtime.agent.plugin.localstorage.route.delegate;
+
+import com.solace.maas.ep.runtime.agent.plugin.route.RouteBundle;
+import com.solace.maas.ep.runtime.agent.plugin.route.delegate.base.MessagingServiceRouteDelegateImpl;
+import org.springframework.stereotype.Component;
+
+import java.util.List;
+
+@Component
+public class DataCollectionFileWriterDelegateImpl extends MessagingServiceRouteDelegateImpl {
+    public DataCollectionFileWriterDelegateImpl() {
+        super("DATA_COLLECTION_FILE_WRITER");
+    }
+
+    @Override
+    public List<RouteBundle> generateRouteList(List<RouteBundle> destinations, List<RouteBundle> recipients,
+                                               String scanType, String messagingServiceId) {
+        return List.of(createRouteBundle(destinations, recipients, scanType, messagingServiceId,
+                "seda:dataCollectionFileWrite", false));
+    }
+}
