+++ service/local-storage-plugin/src/test/java/com/solace/maas/ep/runtime/agent/plugin/localstorage/config/processor/ScanCompleteProcessorImpl.java
@@ -1,0 +1,12 @@
+package com.solace.maas.ep.runtime.agent.plugin.localstorage.config.processor;
+
+import com.solace.maas.ep.runtime.agent.plugin.processor.logging.ScanCompleteProcessor;
+import org.apache.camel.Exchange;
+import org.apache.camel.Processor;
+
+public class ScanCompleteProcessorImpl implements ScanCompleteProcessor, Processor {
+    @Override
+    public void process(Exchange exchange) throws Exception {
+
+    }
+}
