--- service/solace-plugins/src/main/java/com/solace/maas/ep/runtime/agent/plugin/route/aggregation/FileParseAggregationStrategyImpl.java
@@ -1,24 +1,0 @@
-package com.solace.maas.ep.runtime.agent.plugin.route.aggregation;
-
-import com.solace.maas.ep.runtime.agent.plugin.constants.AggregationConstants;
-import com.solace.maas.ep.runtime.agent.plugin.constants.RouteConstants;
-import com.solace.maas.ep.runtime.agent.plugin.jacoco.ExcludeFromJacocoGeneratedReport;
-import org.apache.camel.AggregationStrategy;
-import org.apache.camel.Exchange;
-
-@ExcludeFromJacocoGeneratedReport
-public class FileParseAggregationStrategyImpl implements AggregationStrategy {
-    @Override
-    public Exchange aggregate(Exchange oldExchange, Exchange newExchange) {
-        newExchange.getIn().setHeader(RouteConstants.AGGREGATION_ID,
-                oldExchange.getIn().getHeader(RouteConstants.AGGREGATION_ID));
-        newExchange.getIn().setHeader(AggregationConstants.OBJECT_KEY,
-                oldExchange.getIn().getHeader(AggregationConstants.OBJECT_KEY));
-        newExchange.getIn().setHeader(AggregationConstants.AGGREGATION_SIZE,
-                oldExchange.getIn().getHeader(AggregationConstants.AGGREGATION_SIZE));
-        newExchange.getIn().setHeader(AggregationConstants.COMPLETION_COUNT,
-                oldExchange.getIn().getHeader(AggregationConstants.COMPLETION_COUNT));
-
-        return newExchange;
-    }
-}
