--- service/solace-plugins/src/main/java/com/solace/maas/ep/runtime/agent/plugin/route/delegate/DataCollectionFileWriterDelegateImpl.java
@@ -1,21 +1,0 @@
-package com.solace.maas.ep.runtime.agent.plugin.route.delegate;
-
-import com.solace.maas.ep.runtime.agent.plugin.route.RouteBundle;
-import com.solace.maas.ep.runtime.agent.plugin.route.delegate.base.MessagingServiceRouteDelegateImpl;
-import org.springframework.stereotype.Component;
-
-import java.util.List;
-
-@Component
-public class DataCollectionFileWriterDelegateImpl extends MessagingServiceRouteDelegateImpl {
-    public DataCollectionFileWriterDelegateImpl() {
-        super("DATA_COLLECTION_FILE_WRITER");
-    }
-
-    @Override
-    public List<RouteBundle> generateRouteList(List<RouteBundle> destinations, List<RouteBundle> recipients,
-                                               String scanType, String messagingServiceId) {
-        return List.of(createRouteBundle(destinations, recipients, scanType, messagingServiceId,
-                "seda:dataCollectionFileWrite", false));
-    }
-}
